requirements.txt
Pillow==8.2.0
pywin32==300
